requirements.txt
webdavclient3>=3.14.4
openai>=1.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
PyYAML>=6.0.0